+++ hutool-crypto/src/main/java/cn/hutool/crypto/asymmetric/AsymmetricAlgorithm.java
@@ -13,9 +13,7 @@ public enum AsymmetricAlgorithm {
 	/** RSA算法，此算法用了默认补位方式为RSA/ECB/PKCS1Padding */
 	RSA_ECB_PKCS1("RSA/ECB/PKCS1Padding"), 
 	/** RSA算法，此算法用了RSA/None/NoPadding */
-	RSA_None("RSA/None/NoPadding"), 
-	/** EC（Elliptic Curve）算法 */
-	EC("EC");
+	RSA_None("RSA/None/NoPadding");
 
 	private final String value;